+++ packages/fields/src/components/switch/switch.tsx
@@ -11,7 +11,7 @@ import {
     AttachInternals,
     Watch,
 } from '@stencil/core';
-import { theme } from '@eventstore-ui/theme';
+import { theme } from '@kurrent-ui/theme';
 import type { IconDescription } from '@eventstore-ui/components';
 import { ICON_NAMESPACE } from 'icons/namespace';
 import type { FieldChange } from 'types';